requirements.txt
Flask
openai
PyMuPDF
pytesseract
Pillow
fpdf
